회원가입 이메일인증 마이그레이션 완료 build 오류 건 수정
