+++ next-app/app/character/[nickName]/page.js
@@ -17,7 +17,7 @@ export async function generateMetadata({ params }) {
     title: `${data.profile.NICKNAME} 캐릭터 정보 | LOAGAP`,
     description: `${data.profile.NICKNAME}님의 장비, 보석, 초월, 엘릭서 정보를 확인하세요.`,
     openGraph: {
-      title: `${data.profile.TITLE !== "없음" ? profile.TITLE : ""} ${data.profile.NICKNAME} 캐릭터 정보 | LOAGAP`,
+      title: `${data.profile.TITLE !== "없음" ? data.profile.TITLE : ""} ${data.profile.NICKNAME} 캐릭터 정보 | LOAGAP`,
       description: `${data.profile.SERVER} / ${data.guild.NAME} / Lv.${data.profile.ITEM_LEVEL} / ${data.profile.JOB}·${data.profile.SUBJOB}`,
       url: `https://loagap.com/character/${data.profile.NICKNAME}`,
       images: [
