+++ next-app/app/user/verifyEmail/page.js
@@ -1,33 +1,12 @@
-import BookPage from './bookClient';
+'use client';
 
-async function getBooksPriceData() {
-  const res = await fetch(`${process.env.NEXT_PUBLIC_SERVER_URL}/price/book`, {
-    next: { revalidate: 60 }, // ISR or SSR
-  });
-  if (!res.ok) throw new Error("Failed to fetch");
-  const data = await res.json();
-  return data;
-}
+import VerifyEmailPage from './verifyEmailClient';
+import React, { Suspense } from 'react';
 
-export const metadata = {
-  title: '유각시세 · 유각차트 | LOAGAP',
-  description: '로스트아크의 실시간 유물 각인서 시세, 차트, 가격 변동, 랭킹 정보를 빠르게 확인해보세요.',
-  keywords: 'LOAGAP, 빈틈봇, ,실시간, 유각, 유각시세, 유각차트, 유각가격, 유각랭킹, 유각순위, 유물, 유물 각인서, 각인서, 아드, 원한, 로스트아크',
-  openGraph: {
-    title: '유각시세 | LOAGAP',
-    description: '로스트아크의 실시간 유물 각인서 시세, 차트, 가격 변동, 랭킹 정보를 빠르게 확인해보세요.',
-    url: 'https://loagap.com/price/book',
-    type: 'website',
-  },
-};
-
-export default async function BookPricePage() {
-  const { booksPrice, bookPriceLastUpdate } = await getBooksPriceData();
-
+export default function Page() {
   return (
-    <div className="container-patch">
-      <h2>유각시세 / 유각차트</h2>
-      <BookPage booksPrice={booksPrice} bookLastUpdate={bookPriceLastUpdate} />
-    </div>
+    <Suspense fallback={<div>로딩중...</div>}>
+      <VerifyEmailPage />
+    </Suspense>
   );
 }
